Handle countdown completed in TimerService
We should not rely on the Activity for notification and settings
restoriation. Instead, this should be handled in the TimerService, and
the MainActivity can offer additional notifications.

Ideally, only immediate GUI stuff should remain in MainActivity, and
everything else should be handled elsewehere.

Related issues:
 #36, #22
